Rename maxmod to tracker
